+++ ChangeLog.md
@@ -1,6 +1,11 @@
+### 2024-09-09
+* **Hello.MediatR.Endpoint (2.0.1)**
+  * Upgraded to NET 8.0 from NET 7.0
+
 ### 2021-04-18
-* **Hello.MediatR.Endpoint (1.0.0)**
+* **Hello.MediatR.Endpoint (1.0.1)**
   * Created
+
 ### 2023-03-28
-* **Hello.MediatR.Endpoint (2.0.0)**
-  * Upgrade to NET 7.0 from NET 5.0
+* **Hello.MediatR.Endpoint (1.0.0)**
+  * Upgraded to NET 7.0 from NET 5.0